==> docs/notebooks/08 Features -- TFIDF with Gensim.ipynb <==
from setup_notebooks import *
%matplotlib inline

%matplotlib inline
display(HTML("<style>.container { width:100% !important; }</style>"))
pd.set_option('display.max_rows', 12)
pd.set_option('display.max_columns', 200)

from gensim.models import TfidfModel, LsiModel
from gensim.corpora import Dictionary
from collections import OrderedDict

Load previously cleaned data

dates = pd.read_csv(os.path.join(DATA_PATH, 'datetimes.csv.gz'), engine='python')
nums = pd.read_csv(os.path.join(DATA_PATH, 'numbers.csv.gz'), engine='python')
df = pd.read_csv(os.path.join(DATA_PATH, 'text.csv.gz'))
df.tokens

d = Dictionary.from_documents(([str(s) for s in row]for row in df.tokens))

df.tokens.iloc[0]

# one way to fix this
df.tokens = df.tokens.apply(eval)

### When we said "QUOTE_NONNUMERIC" we didn't mean **ALL** nonnumeric fields ;)

So we can recreate the token lists usint split() again

df['tokens'] = df.txt.str.split()
df.tokens

### That's more like it, our tokens are now lists of strings not stringified lists of strings ;)

df.tokens.values[0:3]

d = Dictionary.from_documents(df.tokens)
d

tfidf = TfidfModel(d)

*Hint-Hint:* `gensim` is sprinting this week at PyCon!

TfidfModel?

TfidfModel(df.txt)

TfidfModel(df.tokens)

TfidfModel((d.doc2bow(tokens) for tokens in df.tokens))

But there's a simpler way.  
We already have a vocabulary  
with term and document frequencies in a matrix...  

pd.Series(d.dfs)

pd.Series(d.iteritems())

OK, now I get it  

- `document` is a list of strings (ordered sequence of tokens)  
- `bow` or [bag of words] is a list of `Counter`-like mappings between word IDs and their count in each document
- `TfidfModel` is a transformation from a BOW into a BORF,  a "bag of relative frequencies"  

TFIDF = BORF = term frequencies normalized by document occurence counts


pd.Series(d.doc2bow(toks) for toks in df.tokens[:6])

Did it assign 0 to the first word it found?  
Sort-of...  

d.token2id['python']

d.token2id['Python']

d.token2id['you']

d[1]  # guesses anyone?

tfidf = TfidfModel(dictionary=d)
tfidf

dfs = pd.Series(OrderedDict(sorted([(d.id2token[i], numdocs) for (i, numdocs) in tfidf.dfs.items()])))
dfs

dfs.iloc[4000:4030]

tfidf.num_docs

tfidf.num_nnz

tfidf.save(os.path.join(DATA_PATH, 'tfidf'))

tfidf2 = TfidfModel.load(os.path.join(DATA_PATH, 'tfidf'))

tfidf2.num_nnz
