==> docs/notebooks/10 Learning -- Predicting Impact.ipynb <==
from setup import *
import sys
if DATA_PATH not in sys.path: sys.path.append(DATA_PATH)
from constants import *

%matplotlib inline
display(HTML("<style>.container { width:100% !important; }</style>"))
pd.set_option('display.max_rows', 6)
pd.set_option('display.max_columns', 300)
pd.set_option('precision', 2)
%pprint

from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

lsi = LsiModel.load(os.path.join(DATA_PATH, 'lsi'))
lsi4 = LsiModel.load(os.path.join(DATA_PATH, 'lsi4'))
lsi.num_topics
lsi.show_topics(1,10)

# favorites
nums = pd.read_csv(os.path.join(DATA_PATH, 'numbers.csv.gz'))
dates = pd.read_csv(os.path.join(DATA_PATH, 'datetimes.csv.gz'))

vocab = Dictionary(df.txt.str.split())

tfidf = TfidfModel(id2word=vocab, dictionary=vocab)

lsi = LsiModel.load(os.path.join(DATA_PATH, 'lsi'))

# Bags of words
df = pd.DataFrame.from_csv(os.path.join(DATA_PATH, 'text.csv.gz'))
bows = pd.Series(vocab.doc2bow(toks) for toks in df.txt.str.split())

topics = lsi[tfidf[bows]]
topics = pd.DataFrame((dict(top) for top in topics), index=df.index)
print(topics.isnull().sum())
topics.fillna(0, inplace=True)
print(topics.isnull().sum())

topics.copy().round(2)

When I first ran this, my dataframes weren't "aligned".  
So it's very important to check your datasets after every load.  
The correspondence between dates and topics and numerical features is critical for training!  

print(len(dates))
print(len(topics))
print(len(nums))

sum(nums.index == dates.index) == len(dates)

sum(nums.index == topics.index) == len(dates)

disc = LinearDiscriminantAnalysis()
disc

category = (np.ceil(nums.favorite_count ** .13)).astype(np.int8)

disc = LinearDiscriminantAnalysis().fit(topics, category)

predicted_favorites = disc.predict(topics)
predicted_favorites[:100]

np.sum(predicted_favorites > 0)

### Wow!  
DiscriminantAnalysis is **VERY** discriminating!

np.sum(nums.favorite_count >= 1)

But not in a good way.  
10x more true favorites than predicted. 
Our unbalanced training set makes it easy for the judge to be tough.  
Let's mellow our judge a bit...  

results = pd.DataFrame()
results['predicted'] = predicted_favorites
results['truth'] = pd.Series(nums.favorite_count >= 1)
conf = Confusion(results)
conf

results.predicted.corr(results.truth)

conf.stats_dict

High accuracy, but low MCC (correlation)

Balance the training?  
Get rid of some negatives?  
**Accentuate the positive?** <-- give this a try yourself

pos = np.array(nums.favorite_count >= 1)
neg = ~pos
portion_pos = float(sum(pos)) / len(nums)
mask = ((np.random.binomial(1, portion_pos, size=len(nums)).astype(bool) & neg) | pos)
disc = LinearDiscriminantAnalysis().fit(topics[mask], (nums.favorite_count[mask] >= 1))
print(sum(mask))
print(sum(pos) * 2)

results = pd.DataFrame()
results['predicted'] = disc.predict(topics.values)
results['truth'] = nums.favorite_count.values >= 1
conf = Confusion(results)
conf

results.predicted.corr(results.truth)

conf.stats_dict

So let's add some more negative examples back in.  
50x imbalance is defintiely misleading.  
But 2-5x imbalance is probably OK.  

portion_neg = 3 * portion_pos
mask = ((np.random.binomial(1, portion_neg, size=len(nums)).astype(bool) & neg) | pos)
disc = LinearDiscriminantAnalysis().fit(topics[mask], nums.favorite_count[mask] >=1 )
print(sum(mask))
print(sum(pos) * 2)

results = pd.DataFrame()
results['predicted'] = disc.predict(topics.values)
results['truth'] = nums.favorite_count.values > 0
conf = Confusion(results)
conf

At least the confusion matrix looks balanced now

results.predicted.corr(results.truth)

Should have known, imbalance doesn't help...

portion_neg = 2 * portion_pos
mask = ((np.random.binomial(1, portion_neg, size=len(nums)).astype(bool) & neg) | pos)
disc = LinearDiscriminantAnalysis().fit(topics.values[mask], (nums.favorite_count.values > 0)[mask])
print(sum(mask))
print(sum(pos) * 2)

results = pd.DataFrame()
results['predicted'] = disc.predict(topics.values)
results['truth'] = nums.favorite_count.values > 0
conf = Confusion(results)
conf

results.predicted.corr(results.truth)

So it looks like 38% correlation is all we can squeeze out of this simple model  
Next up... adding number of followers and other features
